Add files via Upload
Program using decorator feature to add up elements of two list in and make it as new list containing the summed element.
Note : here adding up list means let A=[1,2,3,4] and B=[2,3,4,5] and output should be as new list containing sum of elements parallel to the positional value in given respective order of list. 
Note : Do not add as A+B , both A and B should have dynamic input. adding using + operator will concatenate the elements of list from the last position of preceding list.